backend/requirements.txt
Django>=5.2,<6
djangorestframework>=3.14
psycopg2-binary>=2.9
PyJWT>=2.8
bcrypt>=4.1
django-cors-headers>=4.3
requests>=2.31
beautifulsoup4>=4.12
boto3>=1.34
cloudinary>=1.38
python-dotenv>=1.0
gunicorn>=22.0
msgpack>=1.1
websockets>=12,<16
